% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat.R
\name{cheat}
\alias{cheat}
\title{cheat}
\usage{
cheat(number)
}
\arguments{
\item{number}{A data set object}
}
\value{
A string containing the correct solution of a given exercise.
}
\description{
Provides the correct solution of exercises 1,2,15 & 16 of assignment 3.1
}
