% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind_me.R
\name{remind_me}
\alias{remind_me}
\title{remind_me}
\usage{
remind_me()
}
\value{
A list containing the date and the mode of the exam as well as the teacher names.
}
\description{
Provides information about the exam for the course "Programming in Psychological Science"
}
